requirements.txt
# Core Framework
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
pydantic>=2.12.5
pydantic-settings>=2.12.0
python-multipart>=0.0.6
jinja2>=3.1.3
itsdangerous>=2.1.2

# LLM Integration
langchain>=1.2.7
langchain-anthropic>=1.3.1
langchain-openai>=1.1.7
langchain-google-genai>=4.2.0
anthropic>=0.76.0
openai>=2.14.0
google-genai>=1.60.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Session middleware
starlette>=0.46.0

# Async Support
httpx>=0.26.0

# Document Parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Web Scraping (optional - falls back to httpx)
# crawl4ai>=0.3.0

# Utilities
python-dotenv>=1.0.0
